requirements.txt
Django==3.0.6
djangorestframework==3.11.0
pillow
psycopg2